Part B. Agent Application/requirements.txt
google-adk
pysqlite3-binary
pdfplumber
google-cloud-aiplatform
cloudpickle==3.1.1
pydantic==2.10.6
pytest
overrides
scikit-learn
reportlab
google-cloud-storage
deprecated
wikipedia